Product (#47)
* fix product specification

* Show variant price

* price  display and authentication fix

* show variant price

* Show min price as default

* hide 0 price

* show variant price in cart

* show variant price on cart and order

+++ sm-shop/src/main/java/com/salesmanager/shop/store/controller/shoppingCart/facade/ShoppingCartFacadeImpl.java
@@ -165,7 +165,7 @@ public class ShoppingCartFacadeImpl
         	Set<com.salesmanager.core.model.shoppingcart.ShoppingCartItem> cartModelItems = cartModel.getLineItems();
         	for(com.salesmanager.core.model.shoppingcart.ShoppingCartItem cartItem : cartModelItems) {
         		if(cartItem.getProduct().getId().longValue()==shoppingCartItem.getProduct().getId().longValue()) {
-        		    if(cartItem.getPriceId()==shoppingCartItem.getPriceId()) {
+        		    if(cartItem.getPriceId().longValue()==shoppingCartItem.getPriceId().longValue()) {
 //        			if(CollectionUtils.isEmpty(cartItem.getAttributes())) {
         				if(!duplicateFound) {
 //        					if(!shoppingCartItem.isProductVirtual()) {
